+++ IsContractBulk.sol
@@ -3,40 +3,34 @@
 pragma solidity ^0.8.4;
 
 contract IsContractBulk {
-    function isContract(address addr) 
-        public
-        view
-        returns (bool)
-    {
+    function isContract(address addr) public view returns (bool) {
         uint256 size;
-        assembly { size := extcodesize(addr) }
+        assembly {
+            size := extcodesize(addr)
+        }
         return size > 0;
     }
-    
-    function removeIsContractBulk(address[] memory addrs) 
+
+    function removeIsContractBulk(address[] memory addrs)
         public
         view
         returns (address[] memory)
     {
-        for (uint i = 0; i<addrs.length; i++)
-        {
-            if (isContract(addrs[i]))
-            {
+        for (uint256 i = 0; i < addrs.length; i++) {
+            if (isContract(addrs[i])) {
                 addrs[i] = address(0x0);
             }
         }
         return addrs;
     }
-    
-    function removeIsNotContractBulk(address[] memory addrs) 
+
+    function removeIsNotContractBulk(address[] memory addrs)
         public
         view
         returns (address[] memory)
     {
-        for (uint i = 0; i<addrs.length; i++)
-        {
-            if (!isContract(addrs[i]))
-            {
+        for (uint256 i = 0; i < addrs.length; i++) {
+            if (!isContract(addrs[i])) {
                 addrs[i] = address(0x0);
             }
         }
